packages/internal-auth/pyproject.toml
[project]
name = "internal-auth"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
